pyproject.toml
[project]
name = "sale-agent"
version = "0.1.0"
description = "WhatsApp visual product search for Shopify stores"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "torchvision>=0.26.0",
    "transformers>=5.5.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "twilio>=9.0.0",
    "python-multipart>=0.0.9",
]
